[frontend] e2e fixes (#14345)
